Convert Request/Response Player Position Type from Int to Float
Converted all int type for
 request/response relating to player position to float.
Also add ability to send float value through request by adding in missing statement/function in GamePacket/GamePacketStream.

+++ Client/Assets/Scripts/Network/NetworkManager.cs
@@ -51,7 +51,7 @@ public class NetworkManager : MonoBehaviour
     // }
 
     // Request to spawn current client's player
-    public bool RequestSpawnPlayer(int x, int y)
+    public bool RequestSpawnPlayer(float x, float y)
     {
         if (cManager && cManager.IsConnected())
         {
